+++ R/TheilSen.R
@@ -165,6 +165,7 @@
 ##' @param plot Should a plot be produced. \code{FALSE} can be useful when
 ##'   analysing data to extract trend components and plotting them in other
 ##'   ways.
+##' @param silent When \code{FALSE} the function will give updates on trend-fitting progress.
 ##' @param ... Other graphical parameters passed onto \code{cutData}
 ##'   and \code{lattice:xyplot}. For example, \code{TheilSen} passes
 ##'   the option \code{hemisphere = "southern"} on to \code{cutData}
@@ -257,7 +258,7 @@ TheilSen <- function(mydata, pollutant = "nox", deseason = FALSE,
                      text.col = "darkgreen", slope.text = NULL, cols = NULL,
                      shade = "grey95", auto.text = TRUE,
                      autocor = FALSE, slope.percent = FALSE, date.breaks = 7,
-                     plot = TRUE, ...) {
+                     plot = TRUE, silent = FALSE,  ...) {
 
   ## get rid of R check annoyances
   a <- b <- lower.a <- lower.b <- upper.a <- upper.b <- slope.start <- date.end <- intercept.start <- date.start <- lower.start <- intercept.lower.start <- upper.start <- intercept.upper.start <- NULL
@@ -431,7 +432,7 @@ TheilSen <- function(mydata, pollutant = "nox", deseason = FALSE,
 
     ## now calculate trend, uncertainties etc ###########################
     if (nrow(results) < 6) return(results) ## need enough data to calculate trend
-    MKresults <- MKstats(results$date, results$conc, alpha, autocor)
+    MKresults <- MKstats(results$date, results$conc, alpha, autocor, silent = silent)
 
     ## make sure missing data are put back in for plotting
     results <- merge(all.results, MKresults, by = "date", all = TRUE)
@@ -719,8 +720,8 @@ panel.shade <- function(split.data, start.year, end.year, ylim,
     ))
 }
 
-MKstats <- function(x, y, alpha, autocor) {
-  estimates <- regci(as.numeric(x), y, alpha = alpha, autocor = autocor)$regci
+MKstats <- function(x, y, alpha, autocor, silent) {
+  estimates <- regci(as.numeric(x), y, alpha = alpha, autocor = autocor, pr = silent)$regci
 
   p <- estimates[2, 5]