+++ R/senTheil.R
@@ -42,7 +42,7 @@ regci <- function(x, y, regfun = tsreg, nboot = 599, alpha = 0.05, autocor = aut
     set.seed(2)
   } ## set seed of random number generator so that
   ##             results can be duplicated.
-  if (pr) {
+  if (!pr) {
     print("Taking bootstrap samples. Please wait.")
   }